+++ Projects/App/Sources/Data/Manager/Storage/KeychainManager.swift
@@ -9,51 +9,37 @@
 import Foundation
 import Security
 
-protocol KeychainManagerProtocol {
-    func save(key: String, data: Data) -> OSStatus
-    func load(key: String) -> Data?
-    func delete(key: String)
-}
-
-final class KeychainManager: KeychainManagerProtocol {
+final class KeychainManager {
+//    static let shared = KeychainManager()
     
-    static let shared = KeychainManager()
-    
-    enum KeychainError: Error {
-        case noData
-    }
-    
-    @discardableResult
-    func save(key: String, data: Data) -> OSStatus {
-        let query = [
-            kSecClass as String: kSecClassGenericPassword,
+    static func save(key: String, data: Data) {
+        let query: [String: Any] = [
+            kSecClass as String: kSecClassGenericPassword as String,
             kSecAttrAccount as String: key,
             kSecValueData as String: data
-        ] as [String: Any]
+        ]
 
-        SecItemDelete(query as CFDictionary)
-        return SecItemAdd(query as CFDictionary, nil)
+        SecItemAdd(query as CFDictionary, nil)
     }
-    
-    func load(key: String) -> Data? {
-        let query = [
+
+    static func load(key: String) -> Data? {
+        let query: [String: Any] = [
             kSecClass as String: kSecClassGenericPassword,
             kSecAttrAccount as String: key,
             kSecReturnData as String: kCFBooleanTrue!,
             kSecMatchLimit as String: kSecMatchLimitOne
-        ] as [String: Any]
-        
-        var dataTypeRef: AnyObject? = nil
+        ]
+
+        var dataTypeRef: AnyObject?
         let status: OSStatus = SecItemCopyMatching(query as CFDictionary, &dataTypeRef)
-        
         if status == noErr {
             return dataTypeRef as? Data
         } else {
             return nil
         }
     }
-    
-    func delete(key: String) {
+
+    static func delete(key: String) {
         let query = [
             kSecClass as String: kSecClassGenericPassword as String,
             kSecAttrAccount as String: key
@@ -62,3 +48,55 @@ final class KeychainManager: KeychainManagerProtocol {
         SecItemDelete(query as CFDictionary)
     }
 }
+
+//protocol KeychainManagerProtocol {
+//    func save(key: String, data: Data)
+//    func load(key: String) -> Data?
+//    func delete(key: String)
+//}
+//
+//final class KeychainManager: KeychainManagerProtocol {
+//    
+//    static let shared = KeychainManager()
+//    
+//    enum KeychainError: Error {
+//        case noData
+//    }
+//    
+//    static func save(key: String, data: Data) {
+//        let query: [String: Any] = [
+//            kSecClass as String: kSecClassGenericPassword as String,
+//            kSecAttrAccount as String: key,
+//            kSecValueData as String: data
+//        ]
+//
+//        SecItemAdd(query as CFDictionary, nil)
+//    }
+//    
+//    static func load(key: String) -> Data? {
+//        let query = [
+//            kSecClass as String: kSecClassGenericPassword,
+//            kSecAttrAccount as String: key,
+//            kSecReturnData as String: kCFBooleanTrue!,
+//            kSecMatchLimit as String: kSecMatchLimitOne
+//        ] as [String: Any]
+//        
+//        var dataTypeRef: AnyObject? = nil
+//        let status: OSStatus = SecItemCopyMatching(query as CFDictionary, &dataTypeRef)
+//        
+//        if status == noErr {
+//            return dataTypeRef as? Data
+//        } else {
+//            return nil
+//        }
+//    }
+//    
+//    static func delete(key: String) {
+//        let query = [
+//            kSecClass as String: kSecClassGenericPassword as String,
+//            kSecAttrAccount as String: key
+//        ]
+//
+//        SecItemDelete(query as CFDictionary)
+//    }
+//}
