+++ Projects/App/Sources/Data/Service/ApiService.swift
@@ -96,6 +96,7 @@ final class ApiService {
                     return .failure(NetworkError.decode)
                 }
                 
+                print("🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵성공 \(data)🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵🩵")
                 return .success(data)
                 
             } catch {
