+++ Projects/App/Sources/Presentation/Login/Term/TermView.swift
@@ -10,55 +10,11 @@ import Combine
 import DesignSystem
 import SwiftUI
 
-enum Term: CaseIterable{
-    case term
-    case personalInfo
-    case marketing
-    
-    var title: String {
-        switch self {
-        case .term:
-            return "(필수) 서비스 이용약관 동의"
-        case .personalInfo:
-            return "(필수) 개인정보 처리방침 동의"
-        case .marketing:
-            return "(선택) 마케팅 수신 동의"
-        }
-    }
-}
-
-final class TermViewModel: ObservableObject {
-    @Published var isAllChecked: Bool = false
-    @Published var isTermChecked: Bool = false
-    @Published var isPersonalChecked: Bool = false
-    @Published var isMarketingChecked: Bool = false
-    
-    private var cancellables = [AnyCancellable]()
-    
-    init() {
-        Publishers.CombineLatest3($isTermChecked, $isPersonalChecked, $isMarketingChecked)
-            .sink { [weak self] term, personalInfo, marketing in
-                self?.isAllChecked = term && personalInfo || marketing
-            }
-            .store(in: &cancellables)
-    }
-    
-    func toggleAll() {
-        if isTermChecked && isPersonalChecked && isMarketingChecked {
-            isTermChecked = false
-            isPersonalChecked = false
-            isMarketingChecked = false
-        } else {
-            isTermChecked = true
-            isPersonalChecked = true
-            isMarketingChecked = true
-        }
-    }
-}
-
 struct TermView: View {
+    @EnvironmentObject var router: Router
+    @EnvironmentObject var authViewModel: AuthViewModel
+    
     @StateObject var viewModel = TermViewModel()
-    @ObservedObject var authViewModel: AuthViewModel
     @State private var isTermChecked: Bool = false
     
     var body: some View {
@@ -129,6 +85,6 @@ struct TermView: View {
     }
 }
 
-//#Preview {
-//    TermView(authViewModel: AuthViewModel(authUseCase: SignInUseCase(signInRepoProtocol: SignInRepository())))
-//}
+#Preview {
+    TermView()
+}
