+++ Projects/App/Sources/Presentation/Tabbar+Navigation/TabbarMainView.swift
@@ -28,45 +28,45 @@ struct TabbarMainView: View {
     }
 }
 
-struct AuthenticationView: View {
-    @StateObject private var authViewModel = AuthViewModel(
-        accountUseCase: AccountUseCase(
-            accountRepoProtocol: AccountRepository(
-                apiService: ApiService())
-        ),
-        socialUseCase: SocialUsecase(
-            socialRepoProtocol: SocialRepository()
-        )
-    )
-    
-    @StateObject private var viewModel = TabbarViewModel()
-    
-    var body: some View {
-        VStack {
-            switch authViewModel.authenticationState {
-            case .initial:
-                LoginMainView(authViewModel: authViewModel)
-            case .signIn:
-                VStack(spacing: 0) {
-                    TabView(selection: $viewModel.selected) {
-                        ForEach(Tabbar.allCases, id: \.self) { tab in
-                            tab.view
-                        }
-                        .toolbarBackground(.hidden, for: .tabBar)
-                    }
-                    TabbarView(viewModel: viewModel)
-                }
-            case .nickname:
-                NicknameView(authViewModel: authViewModel)
-            case .term:
-                TermView(authViewModel: authViewModel)
-            case .needToToken:
-                EmptyView()
-            }
-        }
-    }
-}
+//struct AuthenticationView: View {
+//    @StateObject private var authViewModel = AuthViewModel(
+//        accountUseCase: AccountUseCase(
+//            accountRepoProtocol: AccountRepository(
+//                apiService: ApiService())
+//        ),
+//        socialUseCase: SocialUsecase(
+//            socialRepoProtocol: SocialRepository()
+//        )
+//    )
+//    
+//    @StateObject private var viewModel = TabbarViewModel()
+//    
+//    var body: some View {
+//        VStack {
+//            switch authViewModel.authenticationState {
+//            case .initial:
+//                LoginMainView(authViewModel: authViewModel)
+//            case .signIn:
+//                VStack(spacing: 0) {
+//                    TabView(selection: $viewModel.selected) {
+//                        ForEach(Tabbar.allCases, id: \.self) { tab in
+//                            tab.view
+//                        }
+//                        .toolbarBackground(.hidden, for: .tabBar)
+//                    }
+//                    TabbarView(viewModel: viewModel)
+//                }
+//            case .nickname:
+//                NicknameView(authViewModel: authViewModel)
+//            case .term:
+//                TermView(authViewModel: authViewModel)
+//            case .needToToken:
+//                EmptyView()
+//            }
+//        }
+//    }
+//}
 
-#Preview {
-    AuthenticationView()
-}
+//#Preview {
+//    AuthenticationView()
+//}
